fix: correct the transaction schema

+++ src/modules/transaction/transaction.schema.ts
@@ -10,7 +10,7 @@ const transactionCore = z.object({
   coin: z.nativeEnum(TransactionCoin, { required_error: 'Coin is required' }),
   transactionType: z.nativeEnum(TransactionType, { required_error: 'Transaction type is required' }),
   amount: z.number({ required_error: 'Amount is required' }).int().min(1),
-  coinAmount: z.number({ required_error: 'Coin Equivalent is required' }).int().nonnegative(),
+  coinAmount: z.number({ required_error: 'Coin Equivalent is required' }).nonnegative(),
   network: z.string().optional(),
   transactionHash: z.string().optional(),
   walletAddress: z.string().optional(),
